requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
dnspython==2.6.1
fastapi==0.112.0
h11==0.14.0
idna==3.7
kafka-python==2.0.2
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.5
